Tp3[Cambios en el código de clase Auto]

+++ src/main/java/dominio/Auto.java
@@ -25,7 +25,7 @@ public class Auto {
 
 
     public static Auto instanciaAuto(int id, String marca, String modelo, String color, String cilindrada,String puerta, String patente, String anio) throws ExceptionAuto {
-        if (Integer.parseInt(cilindrada) < 0 || Integer.parseInt(puerta) < 0 || Integer.parseInt(anio) < 0)
+        if (Double.parseDouble(cilindrada)< 0 || Integer.parseInt(puerta) < 0 || Integer.parseInt(anio) < 0)
             throw new ExceptionAutoNegativeAtributte("Error: The atributtes cilindrada,puerta,anio cannot be negative");
         if  (marca.equals("") || modelo.equals("") || cilindrada.equals("") || puerta.equals("") || patente.equals("") || anio.equals("") )
             throw new ExceptionAutoAtributoVacio("Error: Solo el campo color puede estar vacio ");
